Refactored code to organize into packages.
