+++ docs/com/brownian/morse/LatinReceiver.html
@@ -98,7 +98,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>com.brownian.morse.LatinReceiver</li>
+<li>com.brownian.morse.receivers.LatinReceiver</li>
 </ul>
 </li>
 </ul>
@@ -128,7 +128,7 @@ extends java.lang.Object</pre>
 <th class="colOne" scope="col">Constructor and Description</th>
 </tr>
 <tr class="altColor">
-<td class="colOne"><code><span class="memberNameLink"><a href="../../../com/brownian/morse/LatinReceiver.html#LatinReceiver-com.brownian.morse.MorseReceiver-">LatinReceiver</a></span>(<a href="../../../com/brownian/morse/MorseReceiver.html" title="interface in com.brownian.morse">MorseReceiver</a>&nbsp;receiver)</code>
+<td class="colOne"><code><span class="memberNameLink"><a href="../../../com/brownian/morse/LatinReceiver.html#LatinReceiver-com.brownian.morse.receivers.MorseReceiver-">LatinReceiver</a></span>(<a href="../../../com/brownian/morse/MorseReceiver.html" title="interface in com.brownian.morse">MorseReceiver</a>&nbsp;receiver)</code>
 <div class="block">Creates a LatinReceiver that sends <a href="../../../com/brownian/morse/Morse.Symbol.html" title="enum in com.brownian.morse"><code>Morse symbols</code></a> to the given <a href="../../../com/brownian/morse/MorseReceiver.html" title="interface in com.brownian.morse"><code>MorseReceiver</code></a>.</div>
 </td>
 </tr>
