+++ docs/index-files/index-3.html
@@ -76,7 +76,7 @@
 <dd>
 <div class="block">Accepts a character and sends it to a MorseReceiver as an array of <a href="../com/brownian/morse/Morse.Symbol.html" title="enum in com.brownian.morse"><code>Morse symbols</code></a></div>
 </dd>
-<dt><span class="memberNameLink"><a href="../com/brownian/morse/LatinReceiver.html#LatinReceiver-com.brownian.morse.MorseReceiver-">LatinReceiver(MorseReceiver)</a></span> - Constructor for class com.brownian.morse.<a href="../com/brownian/morse/LatinReceiver.html" title="class in com.brownian.morse">LatinReceiver</a></dt>
+<dt><span class="memberNameLink"><a href="../com/brownian/morse/LatinReceiver.html#LatinReceiver-com.brownian.morse.receivers.MorseReceiver-">LatinReceiver(MorseReceiver)</a></span> - Constructor for class com.brownian.morse.<a href="../com/brownian/morse/LatinReceiver.html" title="class in com.brownian.morse">LatinReceiver</a></dt>
 <dd>
 <div class="block">Creates a LatinReceiver that sends <a href="../com/brownian/morse/Morse.Symbol.html" title="enum in com.brownian.morse"><code>Morse symbols</code></a> to the given <a href="../com/brownian/morse/MorseReceiver.html" title="interface in com.brownian.morse"><code>MorseReceiver</code></a>.</div>
 </dd>
